+++ lib/Search/Tests/Unit/Adapter/TolerantParser/source/class_with_method.test
